requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==2.3.0
pandas==2.2.2
numpy==1.26.4
pydantic==2.11.7
python-dotenv==1.0.1
python-multipart==0.0.20
pdfminer.six==20250506
fpdf2==2.8.4
langdetect==1.0.9
camelot-py==0.11.0
pillow==11.2.0
lxml==5.2.2
requests==2.32.3
starlette==0.47.3
